refactor: move helper fns to separate files in utils folder
