+++ src/app/(core)/teachers/programmes/[slug]/units/[unitSlug]/lessons/Components/UnitHeader/UnitHeader.test.tsx
@@ -7,7 +7,7 @@ import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import { setUseUserReturn } from "@/__tests__/__helpers__/mockClerk";
 import { mockLoggedIn } from "@/__tests__/__helpers__/mockUser";
 import teachersUnitOverviewFixture from "@/node-lib/curriculum-api-2023/fixtures/teachersUnitOverview.fixture";
-import { getProgrammeStateForUnit } from "@/context/TeacherBrowseAnalytics/helpers";
+import { getProgrammeStateForUnit } from "@/context/TeacherBrowseAnalytics/utils/getProgrammeState";
 import { TeacherBrowseAnalyticsStoreProvider } from "@/context/TeacherBrowseAnalytics/TeacherBrowseAnalyticsProvider";
 
 const track = {
