+++ src/context/TeacherBrowseAnalytics/utils/getProgrammeState.ts
@@ -1,57 +1,10 @@
 import {
-  LessonPathwayData,
-  ProgrammePathwayData,
-  ProgrammeState,
   ProgrammeStateLesson,
   ProgrammeStateUnit,
-  UnitPathwayData,
-} from "./teacherBrowseAnalytics.types";
+} from "../teacherBrowseAnalytics.types";
 
 import { TeachersLessonOverviewPageData } from "@/node-lib/curriculum-api-2023/queries/teachersLessonOverview/teachersLessonOverview.schema";
-import {
-  ExamBoardValueType,
-  KeyStageTitleValueType,
-  PathwayValueType,
-  TierNameValueType,
-} from "@/browser-lib/avo/Avo";
 import { TeachersUnitOverviewData } from "@/node-lib/curriculum-api-2023/queries/teachersUnitOverview/teachersUnitOverview.schema";
-
-export const getProgrammeAnalyticsProperties = (
-  programmeState: ProgrammeState,
-): ProgrammePathwayData => {
-  return {
-    examBoard: programmeState.examboard?.title as ExamBoardValueType,
-    keyStageSlug: programmeState.keystage.slug,
-    keyStageTitle: programmeState.keystage.title as KeyStageTitleValueType,
-    pathway: programmeState.pathway?.title as PathwayValueType,
-    subjectSlug: programmeState.subject.slug,
-    subjectTitle: programmeState.subject.title,
-    tierName: programmeState.tier?.title as TierNameValueType,
-  };
-};
-
-export const getUnitAnalyticsProperties = (
-  programmeState: ProgrammeStateUnit,
-): UnitPathwayData => {
-  const programmePathwayData = getProgrammeAnalyticsProperties(programmeState);
-  return {
-    ...programmePathwayData,
-    unitName: programmeState.unit.title,
-    unitSlug: programmeState.unit.slug,
-  };
-};
-
-export const getLessonAnalyticsProperties = (
-  programmeState: ProgrammeStateLesson,
-): LessonPathwayData => {
-  const unitPathwayData = getUnitAnalyticsProperties(programmeState);
-  return {
-    ...unitPathwayData,
-    lessonName: programmeState.lesson.title,
-    lessonSlug: programmeState.lesson.slug,
-    lessonReleaseDate: programmeState.lesson.lessonReleaseDate,
-  };
-};
 
 export const getProgrammeStateForUnit = (
   data: TeachersUnitOverviewData,
